# Producto/back-fadebooker/openapi.yaml
openapi: 3.0.0
info:
  title: FadeBooker API
  description: |
    Sistema de gestión de citas para barberías - Marketplace de barbería
    
    **Versión:** 1.11.0 (Refactorización técnica y Fase 3)
    
    **Base URL:** `https://fadebooker-backend-ok.azurewebsites.net/api`
    
    **Cambios principales (v1.11.0):**
    - Autenticación JWT integrada y campo contrasena.
    - Endpoints de Citas (Agendamiento) validados.
    - Health Check para despliegue en Azure.
    - Servicios vinculados directamente con Barberos.
  version: 1.11.0
  contact:
    name: FadeBooker Team
    email: team@fadebooker.com
  license:
    name: MIT
servers:
  - url: https://fadebooker-backend-ok.azurewebsites.net/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
paths:
...
  /citas/disponibilidad:
    get:
      operationId: verificarDisponibilidad
      summary: Verificar disponibilidad de un barbero
      description: |
        Comprueba si un barbero está libre en un rango de tiempo específico.
      parameters:
        - name: idBarbero
          in: query
          required: true
          schema:
            type: integer
        - name: fecha
          in: query
          required: true
          schema:
            type: string
            format: date
          example: "2026-05-20"
        - name: hora
          in: query
          required: true
          schema:
            type: string
          example: "15:30"
        - name: duracion
          in: query
          required: true
          schema:
            type: integer
          description: Duración en minutos
      responses:
        '200':
          description: Resultado de disponibilidad
          content:
            application/json:
              schema:
                type: object
                properties:
                  disponible:
                    type: boolean
      tags:
        - Citas
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...